refactor: 🔨IdentifierModel 파일 분리, 데이터 이름 변경(#120)
